feat: validate manifest defined MinimumVersion for version specific functionality
